+++ contracts/ts/__tests__/BatchProcessMessageAndQuadVoteTally.test.ts
@@ -5,7 +5,7 @@ jest.setTimeout(1200000)
 import * as ethers from 'ethers'
 
 import { genTestAccounts } from '../accounts'
-import { timeTravel } from '../'
+import { timeTravel } from '../../node_modules/etherlime/cli-commands/etherlime-test/time-travel.js'
 import { deployTestContracts } from '../utils'
 
 import { config } from 'maci-config'
