+++ contracts/ts/__tests__/PublishMessage.test.ts
@@ -16,8 +16,8 @@ import {
     Keypair,
 } from 'maci-domainobjs'
 
+import { timeTravel } from '../../node_modules/etherlime/cli-commands/etherlime-test/time-travel.js'
 import {
-    timeTravel,
     genDeployer,
     genTestAccounts,
     deployMaci,
